converted the entire codebase from `char*` (often MacRoman) to UTF8-encoded `std::string`; it won’t yet run (there's a std::string somewhere that’s being initialized with NULL, which causes the process to crash in loading), plus text rendering code is loaded with exception-throwing TODOs as it’s still to be rebuilt for UTF8, but it *compiles successfully* which is no small achievement

+++ Source_Files/Files/extensions.h
@@ -44,7 +44,7 @@ void set_to_default_physics_file(void);
 /* Proceses the entire physics file.. */
 void import_definition_structures(void);
 
-void *get_network_physics_buffer(int32 *physics_length);
+void *get_network_physics_buffer(int64_t *physics_length);
 void process_network_physics_model(void *data);
 
 uint32_t get_physics_file_checksum();
